# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "mafia",
        "contact": {
            "name": "A2-LAB",
            "url": "http://a2-lab.com/"
        },
        "version": "1"
    },
    "host": "mafia.a2-lab.com",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
...
        "/api/v1/balance/": {
            "get": {
                "tags": [
                    "balance"
                ],
                "summary": "Get profile balance",
                "operationId": "balance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, balance value provided.",
                        "schema": {
                            "type": "intefer",
                            "example": 1234.990000000000009094947017729282379150390625
                        }
                    },
                    "401": {
                        "description": "Wrong token."
                    }
                }
            }
        },
        "/api/v1/balance/details": {
            "get": {
                "tags": [
                    "balance"
                ],
                "summary": "Get details profile balance.",
                "operationId": "balance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, balance value provided.",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/BalanceDetails"
                        }
                    },
                    "401": {
                        "description": "Wrong token."
                    }
                }
            }
        },
        "/api/v1/balance/referral": {
            "get": {
                "tags": [
                    "balance"
                ],
                "summary": "Get referal profile balance.",
                "operationId": "balance",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, balance value provided.",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/ReferralBalance"
                        }
                    },
                    "401": {
                        "description": "Wrong token."
                    }
                }
            }
        },
...
        "/api/v1/games/get": {
            "get": {
                "tags": [
                    "games"
                ],
                "summary": "Get games",
                "operationId": "Get games",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, api status provided",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Game"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/games/update": {
            "get": {
                "tags": [
                    "games"
                ],
                "summary": "Update the game",
                "operationId": "Update the game",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "token",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id of game",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "score",
                        "in": "query",
                        "description": "score",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation, api status provided"
                    },
                    "437": {
                        "description": "This game isn't available now"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Game": {
            "required": [
                "id",
                "title",
                "pivot"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "pivot": {
                    "$ref": "#/definitions/Pivot_g"
                }
            }
        },
        "Pivot_g": {
            "required": [
                "today_times",
                "limit",
                "times",
                "earned",
                "best_score",
                "is_available"
            ],
            "properties": {
                "today_views": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "times": {
                    "type": "integer"
                },
                "earned": {
                    "type": "number"
                },
                "best_score": {
                    "type": "integer"
                },
                "is_available": {
                    "type": "boolean"
                }
            }
        },
...
        "BalanceDetails": {
            "properties": {
                "balance": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                },
                "paid": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                }
            }
        },
        "ReferralBalance": {
            "properties": {
                "referrals": {
                    "type": "object",
                    "$ref": "#/definitions/Referrals"
                },
                "balance": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                },
                "award": {
                    "type": "number",
                    "format": "float",
                    "example": 20
                },
                "paid": {
                    "type": "number",
                    "format": "float",
                    "example": 1234.55999999999994543031789362430572509765625
                }
            }
        },
        "Referrals": {
            "required": [
                "current",
                "max",
                "percent"
            ],
            "properties": {
                "level": {
                    "type": "integer",
                    "example": 2
                },
                "current": {
                    "type": "integer",
                    "example": 2
                },
                "max": {
                    "type": "integer",
                    "example": 3
                },
                "percent": {
                    "type": "integer",
                    "example": 60
                }
            }
        },
...
    },
...
}